+++ application/controllers/SK_Penghasilan.php
@@ -19,6 +19,9 @@ class SK_Penghasilan extends CI_Controller
 	// Fungsi untuk Petugas
 	public function index()
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Daftar Pengajuan Surat Keterangan penghasilan';
 		$data['permohonan_list'] = $this->M_SK_Penghasilan->get_all_permohonan();
 		$this->load->view('admin/sk_Penghasilan/petugas/index', $data);
@@ -26,6 +29,9 @@ class SK_Penghasilan extends CI_Controller
 
 	public function tambah()
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Tambah Pengajuan Surat Keterangan penghasilan';
 		$data['penduduk_list'] = $this->M_Penduduk->get_all_penduduk();
 
@@ -61,6 +67,9 @@ class SK_Penghasilan extends CI_Controller
 
 	public function edit($id)
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Edit Pengajuan Surat Penghasilan';
 		$data['permohonan'] = $this->M_SK_Penghasilan->get_permohonan_by_id($id);
 		$data['penduduk_list'] = $this->M_Penduduk->get_all_penduduk();
@@ -95,6 +104,9 @@ class SK_Penghasilan extends CI_Controller
 
 	public function hapus($id)
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		if ($this->M_SK_Penghasilan->delete_permohonan($id)) {
 			$this->session->set_flashdata('success', 'Pengajuan berhasil dihapus.');
 		} else {
@@ -105,6 +117,9 @@ class SK_Penghasilan extends CI_Controller
 
 	public function hapus_admin($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		if ($this->M_SK_Penghasilan->delete_permohonan($id)) {
 			$this->session->set_flashdata('success', 'Pengajuan berhasil dihapus.');
 		} else {
@@ -116,6 +131,9 @@ class SK_Penghasilan extends CI_Controller
 	// Fungsi untuk Admin
 	public function admin_list()
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Daftar Permohonan Surat Keterangan Penghasilan';
 		$data['permohonan_list'] = $this->M_SK_Penghasilan->get_all_permohonan();
 		$this->load->view('admin/sk_Penghasilan/admin/index', $data);
@@ -123,6 +141,9 @@ class SK_Penghasilan extends CI_Controller
 
 	public function acc_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data = ['status' => 'Disetujui'];
 		$this->M_SK_Penghasilan->update_permohonan($id, $data);
 		$this->session->set_flashdata('success', 'Permohonan berhasil disetujui.');
@@ -131,6 +152,9 @@ class SK_Penghasilan extends CI_Controller
 
 	public function tolak_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Alasan Penolakan';
 		$data['permohonan'] = $this->M_SK_Penghasilan->get_permohonan_by_id($id);
 
@@ -156,6 +180,9 @@ class SK_Penghasilan extends CI_Controller
 
 	public function cetak_surat($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$permohonan = $this->M_SK_Penghasilan->get_permohonan_by_id($id);
 		$data['permohonan'] = $permohonan;
 
@@ -165,6 +192,9 @@ class SK_Penghasilan extends CI_Controller
 
 	public function cancel_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		// Mengubah status permohonan menjadi 'Menunggu'
 		$this->M_SK_Penghasilan->update_status($id, 'Menunggu');
