+++ application/controllers/SK_Usaha.php
@@ -19,6 +19,9 @@ class SK_Usaha extends CI_Controller
 	// Fungsi untuk Petugas
 	public function index()
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Daftar Pengajuan Surat Keterangan Usaha';
 		$data['permohonan_list'] = $this->M_SK_Usaha->get_all_permohonan();
 		$this->load->view('admin/sk_usaha/petugas/index', $data);
@@ -26,6 +29,9 @@ class SK_Usaha extends CI_Controller
 
 	public function tambah()
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Tambah Pengajuan Surat Keterangan Usaha';
 		$data['penduduk_list'] = $this->M_Penduduk->get_all_penduduk();
 
@@ -58,6 +64,9 @@ class SK_Usaha extends CI_Controller
 
 	public function edit($id)
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Edit Pengajuan Surat Usaha';
 		$data['permohonan'] = $this->M_SK_Usaha->get_permohonan_by_id($id);
 		$data['penduduk_list'] = $this->M_Penduduk->get_all_penduduk();
@@ -89,6 +98,9 @@ class SK_Usaha extends CI_Controller
 
 	public function hapus($id)
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		if ($this->M_SK_Usaha->delete_permohonan($id)) {
 			$this->session->set_flashdata('success', 'Pengajuan berhasil dihapus.');
 		} else {
@@ -99,6 +111,9 @@ class SK_Usaha extends CI_Controller
 
 	public function hapus_admin($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		if ($this->M_SK_Usaha->delete_permohonan($id)) {
 			$this->session->set_flashdata('success', 'Pengajuan berhasil dihapus.');
 		} else {
@@ -110,6 +125,9 @@ class SK_Usaha extends CI_Controller
 	// Fungsi untuk Admin
 	public function admin_list()
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Daftar Permohonan Surat Keterangan Usaha';
 		$data['permohonan_list'] = $this->M_SK_Usaha->get_all_permohonan();
 		$this->load->view('admin/sk_usaha/admin/index', $data);
@@ -117,6 +135,9 @@ class SK_Usaha extends CI_Controller
 
 	public function acc_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data = ['status' => 'Disetujui'];
 		$this->M_SK_Usaha->update_permohonan($id, $data);
 		$this->session->set_flashdata('success', 'Permohonan berhasil disetujui.');
@@ -125,6 +146,9 @@ class SK_Usaha extends CI_Controller
 
 	public function tolak_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Alasan Penolakan';
 		$data['permohonan'] = $this->M_SK_Usaha->get_permohonan_by_id($id);
 
@@ -150,6 +174,9 @@ class SK_Usaha extends CI_Controller
 
 	public function cetak_surat($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$permohonan = $this->M_SK_Usaha->get_permohonan_by_id($id);
 		$data['permohonan'] = $permohonan;
 
@@ -159,6 +186,9 @@ class SK_Usaha extends CI_Controller
 
 	public function cancel_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		// Mengubah status permohonan menjadi 'Menunggu'
 		$this->M_SK_Usaha->update_status($id, 'Menunggu');
