+++ application/controllers/Surat_Tugas.php
@@ -19,6 +19,9 @@ class Surat_Tugas extends CI_Controller
 	// Fungsi untuk Petugas
 	public function index()
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Daftar Pengajuan Surat Tugas';
 		$data['permohonan_list'] = $this->M_Surat_Tugas->get_all_permohonan();
 		$this->load->view('admin/surat_tugas/petugas/index', $data);
@@ -26,6 +29,9 @@ class Surat_Tugas extends CI_Controller
 
 	public function tambah()
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Tambah Pengajuan Surat Tugas';
 		$data['penduduk_list'] = $this->M_Penduduk->get_all_penduduk();
 
@@ -56,6 +62,9 @@ class Surat_Tugas extends CI_Controller
 
 	public function edit($id)
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Edit Pengajuan Surat Tugas';
 		$data['permohonan'] = $this->M_Surat_Tugas->get_permohonan_by_id($id);
 		$data['penduduk_list'] = $this->M_Penduduk->get_all_penduduk();
@@ -83,6 +92,9 @@ class Surat_Tugas extends CI_Controller
 
 	public function hapus($id)
 	{
+		if ($this->session->userdata('level') === 'admin') {
+			redirect('dashboard');
+		}
 		if ($this->M_Surat_Tugas->delete_permohonan($id)) {
 			$this->session->set_flashdata('success', 'Pengajuan berhasil dihapus.');
 		} else {
@@ -93,6 +105,10 @@ class Surat_Tugas extends CI_Controller
 
 	public function hapus_admin($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
+
 		if ($this->M_Surat_Tugas->delete_permohonan($id)) {
 			$this->session->set_flashdata('success', 'Pengajuan berhasil dihapus.');
 		} else {
@@ -104,6 +120,9 @@ class Surat_Tugas extends CI_Controller
 	// Fungsi untuk Admin
 	public function admin_list()
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Daftar Permohonan Surat Tugas';
 		$data['permohonan_list'] = $this->M_Surat_Tugas->get_all_permohonan();
 		$this->load->view('admin/surat_tugas/admin/index', $data);
@@ -111,6 +130,9 @@ class Surat_Tugas extends CI_Controller
 
 	public function acc_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data = ['status' => 'Disetujui'];
 		$this->M_Surat_Tugas->update_permohonan($id, $data);
 		$this->session->set_flashdata('success', 'Permohonan berhasil disetujui.');
@@ -119,6 +141,9 @@ class Surat_Tugas extends CI_Controller
 
 	public function tolak_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$data['title'] = 'Alasan Penolakan';
 		$data['permohonan'] = $this->M_Surat_Tugas->get_permohonan_by_id($id);
 
@@ -144,6 +169,9 @@ class Surat_Tugas extends CI_Controller
 
 	public function cetak_surat($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		$permohonan = $this->M_Surat_Tugas->get_permohonan_by_id($id);
 		$data['permohonan'] = $permohonan;
 
@@ -153,6 +181,9 @@ class Surat_Tugas extends CI_Controller
 
 	public function cancel_permohonan($id)
 	{
+		if ($this->session->userdata('level') === 'pegawai') {
+			redirect('dashboard');
+		}
 		// Mengubah status permohonan menjadi 'Menunggu'
 		$this->M_Surat_Tugas->update_status($id, 'Menunggu');
